+++ postmaster/apiv1/admins.py
@@ -7,8 +7,10 @@ an admin to create, delete, and update admins
 
 from flask import request
 from flask_login import login_required, current_user
+import pyqrcode
+from StringIO import StringIO
 from postmaster import db
-from postmaster.models import Admins, Configs
+from postmaster.models import Admins
 from postmaster.utils import json_logger, clear_lockout_fields_on_user
 from postmaster.decorators import json_wrap, paginate
 from postmaster.errors import ValidationError, GenericError
@@ -147,3 +149,115 @@ def unlock_admin(admin_id):
     admin = Admins.query.get_or_404(admin_id)
     clear_lockout_fields_on_user(admin.username)
     return {}, 200
+
+
+@apiv1.route('/admins/<int:admin_id>/twofactor', methods=['GET'])
+@login_required
+@json_wrap
+def twofactor_status(admin_id):
+    """ Returns if 2 factor is enabled or not
+
+        This information is in the main user route,
+        but I added it here as a stub for the URI.
+    """
+    admin = Admins.query.get_or_404(admin_id)
+    return dict(enabled=admin.otp_active)
+
+
+@apiv1.route('/admins/<int:admin_id>/twofactor', methods=['PUT'])
+@login_required
+@json_wrap
+def twofactor_disable(admin_id):
+    """ Disable 2 factor using API.
+
+        Enabling 2 factor from this route is not possible.
+    """
+    admin = Admins.query.get_or_404(admin_id)
+    status = request.get_json(force=True).get('enabled')
+    if status:
+        if status.lower() == "false":
+            admin.otp_active = False
+            try:
+                db.session.add(admin)
+                db.session.commit()
+            except ValidationError as e:
+                raise e
+            except Exception as e:
+                db.session.rollback()
+                json_logger(
+                    'error', current_user.username,
+                    'The following error occurred in twofactor_disable: {0}'.format(str(e)))
+                raise GenericError('The administrator could not be updated')
+            return {}, 200
+        elif status.lower() == "true":
+            raise GenericError("Cannot enable 2 factor from this route - see docs")
+    raise GenericError("An invalid parameter was supplied")
+
+
+@apiv1.route('/admins/<int:admin_id>/twofactor/qrcode', methods=['GET'])
+@login_required
+def qrcode(admin_id):
+    """ Presents the user with a QR code to scan to setup 2 factor authentication
+    """
+    # render qrcode for FreeTOTP
+    admin = Admins.query.get_or_404(admin_id)
+    if admin.id != current_user.id:
+        raise GenericError('You may not view other admin\'s QR codes')
+    if admin.otp_active:
+        return ('', 204)
+    admin.generate_otp_secret()
+    try:
+        db.session.add(admin)
+        db.session.commit()
+    except ValidationError as e:
+        raise e
+    except Exception as e:
+        db.session.rollback()
+        json_logger(
+            'error', current_user.username,
+            'The following error occurred in qrcode: {0}'.format(str(e)))
+        raise GenericError('The administrator could not be updated')
+    url = pyqrcode.create(admin.get_totp_uri())
+    stream = StringIO()
+    url.svg(stream, scale=5)
+    return stream.getvalue().encode('utf-8'), 200, {
+        'Content-Type': 'image/svg+xml',
+        'Cache-Control': 'no-cache, no-store, must-revalidate',
+        'Pragma': 'no-cache',
+        'Expires': '0'}
+
+
+@apiv1.route('/admins/<int:admin_id>/twofactor/verify', methods=['POST'])
+@login_required
+@json_wrap
+def verify_qrcode(admin_id):
+    """ Verifies if the 2 factor token provided is correct
+
+        This will enable 2 factor for a user.
+    """
+    admin = Admins.query.get_or_404(admin_id)
+    if request.get_json(force=True).get('code'):
+        if not admin.otp_secret:
+            raise GenericError("The 2 Factor Secret has not been generated yet")
+        if admin.verify_totp(request.get_json(force=True).get('code')):
+            if not admin.otp_active:
+                audit_message = 'The administrator "{0}" enabled 2 factor'.format(
+                    admin.username)
+                admin.otp_active = True
+                try:
+                    db.session.add(admin)
+                    db.session.commit()
+                    json_logger('audit', current_user.username, audit_message)
+                except ValidationError as e:
+                    raise e
+                except Exception as e:
+                    db.session.rollback()
+                    json_logger(
+                        'error', current_user.username,
+                        'The following error occurred in verify_qrcode: {0}'.format(str(e)))
+                    raise GenericError('The administrator could not be updated')
+            return {}, 200
+        else:
+            raise GenericError("An invalid code was supplied")
+    else:
+        raise ValidationError("The code was not supplied")